requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
paho-mqtt
sqlalchemy
pytz